Moved dbConfig file into the configs folder
